添加日期时间选择器格式化器 `PickerColumnFormatter`

+++ qfluentwidgets/components/widgets/scroll_area.py
@@ -61,7 +61,7 @@ class SmoothScrollBar(QScrollBar):
         # adjust the duration
         dv = abs(value - self.value())
         if dv < 50:
-            self.ani.setDuration(self.duration * dv / 70)
+            self.ani.setDuration(int(self.duration * dv / 70))
         else:
             self.ani.setDuration(self.duration)
